feature: TestComponent added and Render the table on the basis of data
